clippy: clear all 38 warnings + prune stale task docs
Apply clippy 1.96 suggestions across tfs-rust-content and tfs-rust-core
(redundant closures, manual_contains, doc_lazy_continuation,
empty_line_after_doc_comments, needless_borrow, unnecessary_cast,
manual_is_multiple_of, match_like_matches_macro, while_let_loop,
collapsible_if, nonminimal_bool, question_mark, new_without_default,
only_used_in_recursion, unused variables). Gate test-only helpers
behind #[cfg(test)] and remove truly dead helpers. Allow
too_many_arguments on 5 structural entry points. Remove obsolete
tasks/*.md planning docs (superseded by per-feature subplans).

+++ crates/tfs-rust-core/src/pathfinding.rs
@@ -426,6 +426,7 @@ const TSHORTWAY_UNVISITED_H: i32 = i32::MAX;
 
 /// Prefer cardinal when two relaxations reach the same `waylength` (`cract.cc` strict `<` keeps
 /// first-seen; linked-list expand order can still tie — cardinals match live 772 chase traces).
+#[cfg(test)]
 fn tshortway_should_relax(prev_waylength: u32, new_waylength: u32) -> bool {
     new_waylength < prev_waylength
 }
@@ -572,7 +573,7 @@ impl TShortwaySearch {
             }
 
             // `cract.cc:181-184` — signed sum; negative `Waylength` is valid during reverse expand.
-            let distance = manhattan_dist(neighbor_pos, self.origin) as i32;
+            let distance = manhattan_dist(neighbor_pos, self.origin);
             let heuristic =
                 neighbor_wl + neighbor_wp + (self.min_waypoints as i32) * (distance - 1);
